refactor(query): remove iter module
