+++ rs/index/src/query/iters/and_iter.rs
@@ -1,4 +1,4 @@
-use crate::query::iter::{InvertedIndexIter, Iter};
+use crate::query::iters::{InvertedIndexIter, Iter};
 
 pub struct AndIter<'a> {
     pub iters: Vec<Iter<'a>>,
