+++ rs/index/src/query/mod.rs
@@ -1,3 +1,2 @@
-pub mod iter;
 pub mod iters;
 pub mod planner;
